fix: resolve redirect loop when archived user's session is revoked
The proxy was redirecting /login → /dashboard based on stale session
cookie presence, causing an infinite loop after session revocation.
Move the "already logged in" check into a login page server component
that validates the session properly instead of trusting the cookie.
