+++ src/app/login/page.tsx
@@ -1,78 +1,13 @@
-'use client';
+import { getCurrentUser } from '@/lib/session';
+import { redirect } from 'next/navigation';
+import LoginClient from './page.client';
 
-import { Button } from '@/components/ui/button';
-import { Card } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { signIn } from '@/lib/auth-client';
-import { useRouter } from 'next/navigation';
-import { useState } from 'react';
-import { toast } from 'sonner';
+export default async function LoginPage() {
+  const user = await getCurrentUser();
 
-export default function LoginPage() {
-  const router = useRouter();
-  const [loading, setLoading] = useState(false);
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
+  if (user) {
+    redirect('/dashboard');
+  }
 
-  const handleLogin = async (e: React.FormEvent) => {
-    e.preventDefault();
-    setLoading(true);
-
-    try {
-      const result = await signIn.email({
-        email,
-        password,
-      });
-      if (result.data) {
-        router.push('/dashboard');
-      } else {
-        toast.error(result.error.message || 'Failed to login');
-      }
-    } catch (error: any) {
-      console.error('Login error:', error);
-      toast.error(error.message || 'Failed to login');
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-gray-50 flex items-center justify-center p-4">
-      <Card className="w-full max-w-md p-4">
-        <div className="text-center mb-4">
-          <h1 className="text-2xl font-bold">Translation Helper</h1>
-          <p className="text-gray-600">Manage your translations efficiently</p>
-        </div>
-
-        <form onSubmit={handleLogin} className="space-y-4">
-          <div>
-            <Label htmlFor="login-email">Email</Label>
-            <Input
-              id="login-email"
-              type="email"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              required
-              placeholder="your@email.com"
-            />
-          </div>
-          <div>
-            <Label htmlFor="login-password">Password</Label>
-            <Input
-              id="login-password"
-              type="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              required
-              placeholder="••••••••"
-            />
-          </div>
-          <Button type="submit" className="w-full" disabled={loading}>
-            {loading ? 'Logging in...' : 'Login'}
-          </Button>
-        </form>
-      </Card>
-    </div>
-  );
+  return <LoginClient />;
 }
